+++ android/build.gradle
@@ -6,7 +6,6 @@ buildscript {
 
     dependencies {
         classpath 'com.android.tools.build:gradle:3.2.1'
-        classpath 'com.google.gms:google-services:3.2.1'
     }
 }
 
@@ -15,20 +14,6 @@ allprojects {
         google()
         jcenter()
     }
-
-    configurations.all {
-        resolutionStrategy.force "com.android.support:support-v4:26.1.0"
-    }
-}
-
-// This block encapsulates custom properties and makes them available to all
-// modules in the project.
-ext {
-    // The following are only a few examples of the types of properties you can define.
-    compileSdkVersion = 28
-    // You can also use this to specify versions for dependencies. Having consistent
-    // versions between modules can avoid behavior conflicts.
-    //supportLibVersion = '1.0.0-beta01'
 }
 
 rootProject.buildDir = '../build'
