+++ pubspec.yaml
@@ -19,17 +19,10 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  cloud_firestore: ^0.9.7
-  firebase_auth: ^0.8.1+4
-  google_sign_in: ^4.0.1+1
 
-  flutter_bloc:
-    # path: ../../packages/flutter_bloc
-  equatable: ^0.2.0
-  meta: ^1.1.6
-  font_awesome_flutter: ^8.4.0
-
-
+  # The following adds the Cupertino Icons font to your application.
+  # Use with the CupertinoIcons class for iOS style icons.
+  cupertino_icons: ^0.1.2
 
 dev_dependencies:
   flutter_test:
